back-end/requirements.txt
Flask
flask-restful
mysql-connector
mariadb
psycopg2
